+++ dm/CHANGELOG.md
@@ -4,7 +4,6 @@ All notable changes to this project will be documented in this file.
 
 ## CFT Templates
 
-
 ### XX.12.2019
 
 - Updated internal LB and external LB templates according to backend_service.py.schema change [#476] (https://github.com/GoogleCloudPlatform/cloud-foundation-toolkit/pull/476)
@@ -20,6 +19,11 @@ All notable changes to this project will be documented in this file.
 ### 22.11.2019
 
 - Fixed sharedVPC for GKE use case behaviour in 'project' DM template [#469] (https://github.com/GoogleCloudPlatform/cloud-foundation-toolkit/pull/469)
+
+### 22.11.2019
+
+- Cloud Build Trigger support for Github [#470] (https://github.com/GoogleCloudPlatform/cloud-foundation-toolkit/pull/463)
+- CI example for Github triggering Cloudbuild for PRs
 
 ### 21.11.2019